inline buffers into the repeater module
